refactor: - rename screens and update imports - add horizontal padding for empty search screen
